Attempts to fix sporadic segfault on exit
- Split the python JIT test into multiple files per category (simple, nested, or decompose) so that we can debug easily.

- Explicitly clear the QJIT python cache on exit.
